Fix another edge case with lists!

+++ backend/foodbank_southlondon/api/utils.py
@@ -63,9 +63,9 @@ def overwrite_rows(spreadsheet_id: str, rows: List, index: int = 0) -> None:
     new_row_count = len(rows)
     flask.current_app.logger.debug(f"Overwriting all rows with {new_row_count} new rows in {sheet.spreadsheet.title} ({sheet.url}) ...")
     existing_row_count = sheet.row_count
-    if new_row_count:
+    if rows:
         sheet.update(rows, value_input_option="USER_ENTERED")
-    if existing_row_count > new_row_count:
+    if existing_row_count > 1 and existing_row_count > new_row_count:
         sheet.delete_rows((new_row_count or 1) + 1, existing_row_count)
